Fix error for districts history being empty.
Changed the request to return the history of the last 20 days instead of 3.
Unfortunately the RKI no longer gives workday updates to the districts, only weekly updates every Thursday. This means most of the time the history would be empty.
Changing it to 20 fixes the error, but the data obviously is mostly out of date. Fixing this will require a new API.